Move SplineFontProperties to "layout" target
